Update README to improve format for HTML publication
Enabling GitHub pages support on this repo will allow us to link to the
design specs through http://blog.opensds.io/design-specs/. For the most
part, this doesn't require any change in the way we are currently doing
things.

This change makes minor tweaks to the current format to make the HTML
rendering and navigation a little better.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>